requirements.txt
asgiref==3.4.0
Django==3.2.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
Pillow
djangorestframework
markdown
django-filter